Remove output path mangling
It turns out that this attempt at consistency only applied to some build
systems, and ruined it for others.  In particular, Meson consistently uses
paths relative to the build directory in the compilation database, so doing
nothing here results in prefixes like "../src/whatever.c", which is pretty
enough.

So, simplify the code by just removing this.